enhance: Unbundle /monitor from user_cmds.py
